+++ src/rpm/compressor.rs
@@ -11,12 +11,10 @@ pub enum CompressionType {
     Xz,
 }
 
-// 19 is used here as its 19 for fedora
 impl std::str::FromStr for CompressionType {
     type Err = RPMError;
     fn from_str(raw: &str) -> Result<Self, Self::Err> {
         match raw {
-            "none" => Ok(CompressionType::None),
             "gzip" => Ok(CompressionType::Gzip),
             "zstd" => Ok(CompressionType::Zstd),
             "xz" => Ok(CompressionType::Xz),
@@ -27,25 +25,29 @@ impl std::str::FromStr for CompressionType {
 
 pub enum Compressor {
     None(Vec<u8>),
-    Gzip(libflate::gzip::Encoder<Vec<u8>>),
+    Gzip(flate2::write::GzEncoder<Vec<u8>>),
     Zstd(zstd::stream::Encoder<'static, Vec<u8>>),
     Xz(xz2::write::XzEncoder<Vec<u8>>),
 }
 
-impl TryFrom<CompressionType> for Compressor {
+impl TryFrom<CompressionDetails> for Compressor {
     type Error = RPMError;
 
-    fn try_from(value: CompressionType) -> Result<Self, Self::Error> {
+    fn try_from(value: CompressionDetails) -> Result<Self, Self::Error> {
         match value {
-            CompressionType::None => Ok(Compressor::None(Vec::new())),
-            CompressionType::Gzip => {
-                Ok(Compressor::Gzip(libflate::gzip::Encoder::new(Vec::new())?))
-            }
-            CompressionType::Zstd => Ok(Compressor::Zstd(zstd::stream::Encoder::new(
+            CompressionDetails::None => Ok(Compressor::None(Vec::new())),
+            CompressionDetails::Gzip(level) => Ok(Compressor::Gzip(flate2::write::GzEncoder::new(
                 Vec::new(),
-                19,
+                flate2::Compression::new(level),
+            ))),
+            CompressionDetails::Zstd(level) => Ok(Compressor::Zstd(zstd::stream::Encoder::new(
+                Vec::new(),
+                level,
             )?)),
-            CompressionType::Xz => Ok(Compressor::Xz(xz2::write::XzEncoder::new(Vec::new(), 9))),
+            CompressionDetails::Xz(level) => Ok(Compressor::Xz(xz2::write::XzEncoder::new(
+                Vec::new(),
+                level,
+            ))),
         }
     }
 }
@@ -73,32 +75,45 @@ impl Compressor {
     pub(crate) fn finish_compression(self) -> Result<Vec<u8>, RPMError> {
         match self {
             Compressor::None(data) => Ok(data),
-            Compressor::Gzip(encoder) => Ok(encoder.finish().into_result()?),
+            Compressor::Gzip(encoder) => Ok(encoder.finish()?),
             Compressor::Zstd(encoder) => Ok(encoder.finish()?),
             Compressor::Xz(encoder) => Ok(encoder.finish()?),
         }
     }
+}
 
-    pub(crate) fn get_details(&self) -> Option<CompressionDetails> {
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum CompressionDetails {
+    None,
+    Zstd(i32),
+    Gzip(u32),
+    Xz(u32),
+}
+
+impl CompressionDetails {
+    pub(crate) fn compression_type(&self) -> CompressionType {
         match self {
-            Compressor::None(_) => None,
-            Compressor::Gzip(_) => Some(CompressionDetails {
-                compression_level: "9",
-                compression_name: "gzip",
-            }),
-            Compressor::Zstd(_) => Some(CompressionDetails {
-                compression_level: "19",
-                compression_name: "zstd",
-            }),
-            Compressor::Xz(_) => Some(CompressionDetails {
-                compression_level: "9",
-                compression_name: "xz",
-            }),
+            Self::None => CompressionType::None,
+            Self::Gzip(_) => CompressionType::Gzip,
+            Self::Zstd(_) => CompressionType::Zstd,
+            Self::Xz(_) => CompressionType::Xz,
         }
     }
 }
 
-pub(crate) struct CompressionDetails {
-    pub(crate) compression_level: &'static str,
-    pub(crate) compression_name: &'static str,
+impl Default for CompressionDetails {
+    fn default() -> Self {
+        CompressionDetails::None
+    }
+}
+
+impl From<CompressionType> for CompressionDetails {
+    fn from(value: CompressionType) -> Self {
+        match value {
+            CompressionType::None => CompressionDetails::None,
+            CompressionType::Gzip => CompressionDetails::Gzip(9),
+            CompressionType::Xz => CompressionDetails::Xz(9),
+            CompressionType::Zstd => CompressionDetails::Zstd(19),
+        }
+    }
 }